Rename `HdPubKeyRegistry` to `HdPubKeyCache`
